refactor(sdk-node): drop the BuiltVolumeMount / BuiltPatch prefix; align builder return types
`TlsBuilder.build()` returns `TlsConfig`, `DnsBuilder.build()` returns
`DnsConfig`, `SecretBuilder.build()` returns `SecretEntry` — all the
natural names. `MountBuilder.build()` returned `BuiltVolumeMount` and
`PatchBuilder.build()` returned `BuiltPatch[]`. The prefix was a
collision-avoidance hack: `src/mount.ts` and `src/patch.ts` already
exported `VolumeMount` / `Patch` as TypeScript discriminated unions,
so I renamed the napi-emitted shapes with a `Built` prefix.

The right fix: drop the parallel TS-side unions and let the napi types
own the canonical names.

- native/mount_builder.rs: `js_name = "VolumeMount"` (was
  "BuiltVolumeMount").
- native/patch_builder.rs: `js_name = "Patch"` (was "BuiltPatch").
- src/mount.ts deleted.
- src/patch.ts deleted.
- src/index.ts: `VolumeMount` / `Patch` exports now flow through
  `microsandbox/native` (the auto-generated .d.ts) instead of the
  hand-written TS unions. The runtime helpers `VolumeMountKind` /
  `VolumeMountKinds` / `PatchKind` / `PatchKinds` (kind enums + their
  string-literal arrays) are inlined into index.ts.

Trade-off: the napi-emitted `VolumeMount` / `Patch` interfaces have
`kind: string` (instead of a literal-union) and all variant-specific
fields optional, so TypeScript exhaustive narrowing on `kind` is
weaker than the previous TS unions. Users still get autocomplete on
the field names; explicit narrowing via `m.kind === "named"` works at
runtime. This is the same trade-off other auto-generated napi types
already make (e.g. `TlsConfig`).

Verified: tsc clean, vitest 5 files / 45 tests pass.
